feat(ts-subtitles): наполнить крейт реальной логикой + дедуп монолита (#99)
Был пустой крейт-плейсхолдер. Вынес платформо-независимую subtitle-логику из монолитного
subtitles/commands.rs (read/save/validate/convert/get_info для SRT/VTT/ASS) в ts-subtitles
(сняв #[tauri::command]) + перенёс её тесты.

- ts-subtitles: 5 функций + 26 тестов (cargo test зелёный).
- монолит: subtitles/commands.rs → тонкие Tauri-обёртки (486 → 31 строка); mod.rs
  ре-экспортит ТИПЫ из крейта (не функции — иначе коллизия имён с одноимёнными командами,
  которые регистрируются как crate::subtitles::read_subtitle_file в app_builder).

ts-subtitles теперь настоящий протестированный независимый модуль (был vestigial).
Монолит локально не собирается → компиляция с обёртками верифицируется в CI.
