+++ crates/ts-subtitles/Cargo.toml
@@ -16,3 +16,7 @@ uuid = { version = "1", features = ["v4", "serde"] }
 chrono = { version = "0.4", features = ["serde"] }
 anyhow = "1"
 log = "0.4"
+tokio = { version = "1.52.3", features = ["rt", "macros", "rt-multi-thread"] }
+
+[dev-dependencies]
+tempfile = "3.27.0"
