+++ src-tauri/src/subtitles/mod.rs
@@ -1,5 +1,9 @@
-//! ДЕДУП (#99): логика субтитров — из крейта `ts-subtitles`; монолит хранит только Tauri-commands.
-pub use ts_subtitles::subtitles::*;
+//! Субтитры Timeline Studio.
+//!
+//! ДЕДУП (#99): вся логика (read/save/validate/convert/info для SRT/VTT/ASS) — в крейте
+//! `ts-subtitles`; здесь только тонкие Tauri-обёртки (`commands`). Ре-экспортируем ТИПЫ
+//! из крейта (не функции — иначе коллизия с одноимёнными Tauri-командами).
+pub use ts_subtitles::subtitles::{SubtitleExportOptions, SubtitleImportResult, SubtitleInfo};
 
 pub mod commands;
 pub use commands::*;
